move common api utils and requests strategy related composables to api directory
