+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/wahlscheineService.ts
@@ -1,6 +1,6 @@
 import { Configuration } from "@/api/wls-clients/generated-briefwahl-api";
 import { WahlscheineControllerApi } from "@/api/wls-clients/generated-ergebnismeldung-api";
-import { useCommonApiUtils } from "@/composables/common/commonApiUtils.ts";
+import { useCommonApiUtils } from "@/composables/api/commonApiUtils.ts";
 import { useWahlscheineMapper } from "@/composables/ergebnismeldung/wahlscheineMapper.ts";
 import { useUserNotificationService } from "@/composables/userNotification/userNotificationService.ts";
 import { ERGEBNISMELDUNG_SERVICE_API_URL } from "@/constants.ts";
