+++ wls-gui-wahllokalsystem/src/composables/wahlhandlung/waehlerverzeichnisService.ts
@@ -2,7 +2,7 @@ import type { PflegeWaehlerverzeichnis } from "@/types/wahlbezirk/PflegeWaehlerv
 
 import { WaehlerverzeichnisControllerApi } from "@/api/wls-clients/generated-wahlvorbereitung-api";
 import { Configuration } from "@/api/wls-clients/generated-wahlvorstand-api";
-import { useCommonApiUtils } from "@/composables/common/commonApiUtils.ts";
+import { useCommonApiUtils } from "@/composables/api/commonApiUtils.ts";
 import { useUserNotificationService } from "@/composables/userNotification/userNotificationService.ts";
 import { useWaehlerverzeichnisMapper } from "@/composables/wahlhandlung/waehlerverzeichnisMapper.ts";
 import { WAHLVORBEREITUNG_SERVICE_API_URL } from "@/constants.ts";
